+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 + Add more information in `nachos_cook` and catch correct energy with Gaussian (!22)
 + Add Q-Chem (!23), but only for [CCMAN2](http://www.q-chem.com/qchem-website/manual/qchem51_manual/sect-ccmeth.html) energies
 + Output version and parameters (as requested by Benoît)
++ Add uncertainties to the output (!24)
++ Switch to `pip-tools`
 
 ## Version 0.3